requirements.txt
pygame-ce==2.5.1; python_version >= '3.12'
pytmx==3.32; python_version >= '3.12'
